Add an ability system and Change the variable in the status from int to float.

+++ Assets/3 Scripts/3 Inca/UI/ApplicationIcon.cs
@@ -70,7 +70,7 @@ public class ApplicationIcon : DamagableEnemy
         else spriteRenderer.color = new Color(45f / 255, 45f / 255, 45f / 255, c.a);
     }
 
-    public override int TakeDamage(int attckAmount)
+    public override float TakeDamage(float attckAmount)
     {
         if (!canSelect) return 0;
